Version 4.1 - Auto-Authentication, Code-Cleanup
Version 4.1
- DropboxBrowser now handles authentication and login with Dropbox.
There is no need to check if the user is logged in to Dropbox before
presenting the DropboxBrowser. If the user is not logged in,
DropboxBrowser will prompt the user to do so. If the user opts-out of
login then DropboxBrowser will dismiss itself. However, you may still
handle login operations yourself.
- Updated delegate names
- Added code comments and standardized header definitions